\name{commonsMath-package}
\alias{commonsMath-package}
\alias{commonsMath}
\docType{package}
\title{
\packageTitle{commonsMath}
}
\description{
\packageDescription{commonsMath}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{commonsMath}

See the package vignette for details on how to access the Apache Commons Math
Library using this package.
}
\examples{
if ( interactive() ) vignette("commonsMath")
}

